Stick to using size_t everywhere for grid positions

+++ tools/gantry_db_creator/gantry_db_creator.cc
@@ -42,7 +42,7 @@ main()
                  << "isGreyscale" << false;
 
         cv::Mat frame(imSize, CV_8UC3);
-        for (int x = 0, y = 0; x < static_cast<int>(gridRecorder.sizeX()); ) {
+        for (size_t x = 0, y = 0; x < gridRecorder.sizeX(); ) {
             // Move gantry to next position
             const auto pos = gridRecorder.getPosition({ x, y, 0 });
             gantry.setPosition(pos[0], pos[1], pos[2]);
@@ -63,7 +63,7 @@ main()
 
             // If we haven't finished moving along y, move along one more
             if ((x % 2) == 0) {
-                if (y < static_cast<int>(gridRecorder.size())) {
+                if (y < gridRecorder.size()) {
                     y++;
                     continue;
                 }
